chore(lint): expand ruff ruleset and fix findings
Selects pycodestyle core, pyflakes, isort, pyupgrade, bugbear,
comprehensions, return style, ruff rules, performance, datetime tz,
builtins shadowing and bandit. Bandit try/except-pass, subprocess and
SQL-string audit codes stay off (audited patterns); tests are exempt
from bandit. Fixes ~200 findings: unused unpacked/loop vars, manual
comprehensions, mutable class defaults, dangling asyncio tasks,
assert-based invariants, nosec to noqa conversion, shutil.which for
partial paths, and assorted cleanups.

+++ meshchatx/src/backend/http/routes/path_probe.py
@@ -547,9 +547,7 @@ def register_path_probe_routes(routes, app):
             )
         if timeout_seconds is None:
             reticulum = app.reticulum if hasattr(app, "reticulum") else None
-            timeout_seconds = int(
-                round(path_response_window(destination_hash, reticulum)),
-            )
+            timeout_seconds = round(path_response_window(destination_hash, reticulum))
 
         # Split the budget so path discovery cannot consume the whole timeout.
         path_budget_seconds = max(1, timeout_seconds // 2)
